* module/x86/op_model_athlon.c: Don't poke reserved bits in the counter.   Only rdmsr/wrmsr counters that we've enabled in the NMI handler.   ulong->uint conversion.
This still doesn't get things working 100%, but it's an improvement.
Sometimes it still does the reboot instantly thing, but it's not as
reproducable as it was before these fixes.

Comment out the CTR_WRITE line, and its fine, so there's still something
very fishy going on there.  I'll look at it again when I wake up.